4422: Move DockerImageConfiguration and StackDockerImage back to core
