+++ java/clients/src/test/java/sleeper/clients/deploy/UploadDockerImagesTestBase.java
@@ -20,10 +20,8 @@ import org.junit.jupiter.api.BeforeEach;
 import sleeper.clients.testutil.RunCommandTestHelper;
 import sleeper.clients.util.CommandPipeline;
 import sleeper.clients.util.InMemoryEcrRepositories;
-import sleeper.core.deploy.DockerImageConfiguration;
 import sleeper.core.deploy.LambdaHandler;
 import sleeper.core.deploy.LambdaJar;
-import sleeper.core.deploy.StackDockerImage;
 import sleeper.core.properties.PropertiesDiff;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.validation.OptionalStack;
@@ -32,11 +30,11 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import static sleeper.clients.deploy.StackDockerImage.dockerBuildImage;
+import static sleeper.clients.deploy.StackDockerImage.dockerBuildxImage;
+import static sleeper.clients.deploy.StackDockerImage.emrServerlessImage;
 import static sleeper.clients.util.Command.command;
 import static sleeper.clients.util.CommandPipeline.pipeline;
-import static sleeper.core.deploy.StackDockerImage.dockerBuildImage;
-import static sleeper.core.deploy.StackDockerImage.dockerBuildxImage;
-import static sleeper.core.deploy.StackDockerImage.emrServerlessImage;
 import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.VERSION;
 import static sleeper.core.properties.instance.CommonProperty.ACCOUNT;
 import static sleeper.core.properties.instance.CommonProperty.ID;
